Moved classes to split up optional helpers from core tools.
Core = Piwik, Tracker, TrackMe, QueryParams (+ internal dispatcher implementation)
Extra = Everything else.
