+++ exampleapp/src/main/java/com/piwik/demo/SettingsActivity.java
@@ -16,9 +16,9 @@ import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.EditText;
 
-import org.piwik.sdk.PiwikApplication;
-import org.piwik.sdk.TrackHelper;
 import org.piwik.sdk.dispatcher.Packet;
+import org.piwik.sdk.extra.PiwikApplication;
+import org.piwik.sdk.extra.TrackHelper;
 
 import java.util.ArrayList;
 import java.util.Collections;
