+++ piwik-sdk/src/main/java/org/piwik/sdk/extra/DownloadTracker.java
@@ -1,4 +1,4 @@
-package org.piwik.sdk;
+package org.piwik.sdk.extra;
 
 
 import android.content.SharedPreferences;
@@ -7,6 +7,10 @@ import android.content.pm.PackageManager;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import org.piwik.sdk.Piwik;
+import org.piwik.sdk.QueryParams;
+import org.piwik.sdk.TrackMe;
+import org.piwik.sdk.Tracker;
 import org.piwik.sdk.tools.Checksum;
 
 import java.io.File;
