+++ piwik-sdk/src/test/java/org/piwik/sdk/extra/CustomVariablesTest.java
@@ -1,4 +1,4 @@
-package org.piwik.sdk;
+package org.piwik.sdk.extra;
 
 import org.apache.maven.artifact.ant.shaded.StringUtils;
 import org.json.JSONArray;
